moved functions in src files and imported them into calc.mjs for modularization
