+++ package.json
@@ -2,10 +2,11 @@
   "name": "calclite",
   "version": "1.0.0",
   "description": "CalcLite is a simple command-line calculator built with Node.js. It supports basic arithmetic expressions using +, -, *, and /.",
-  "main": "calc.js",
+  "main": "calc.mjs",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node calc.js"
+    "start": "node calc.mjs"
   },
   "repository": {
     "type": "git",
